+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.141.0] - Not released
+### Changed
+- Migrate to React 19.
+- Migrate to ESLint 9 with flat config.
+
 ## [1.140.2] - 2025-09-20
 ### Fixed
 - The sidebar no longer closes when clicking on group links in the "Recent